Split code into Baker and Processor files
